### 2025-05-17 #8246 @editorialbot
[PRE REVIEW]: RANO2.0-assist: A 3D Slicer Extension for (semi-)automatic Assessment of Response to Glioma Treatment
**Submitting author:** @aaronkujawa (<a href="http://orcid.org/0000-0002-1445-6207">Aaron Kujawa</a>)
**Repository:** https://github.com/cai4cai/rano2.0-assist
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
