### 2025-05-23 4260e56
Do not empty lineNodePairs on Reload so report can still be created
