### 2025-05-23 a52b687
Improved report creation
