+++ RANO/utils/report_creation_utils.py
@@ -1,17 +1,23 @@
 import os
-import shutil
 from datetime import datetime
 import json
 import ScreenCapture
 import numpy as np
+from collections import defaultdict
+
+from reportlab.lib import utils, colors
+from reportlab.lib.enums import TA_CENTER, TA_LEFT
+from reportlab.platypus import SimpleDocTemplate, Table, TableStyle, Paragraph, Spacer, KeepTogether, Image, Preformatted
+from reportlab.lib.pagesizes import letter
+from reportlab.lib.units import inch
+from reportlab.lib.styles import getSampleStyleSheet, ParagraphStyle
 
 import slicer
 import qt
-from utils.config import reports_path
 
+from utils.config import reports_path, debug
 import utils.measurements2D_utils as measurements2D_utils
 import utils.segmentation_utils as segmentation_utils
-
 from utils.rano_utils import find_closest_plane
 
 
@@ -77,6 +83,7 @@ class ReportCreationMixin:
         self.table_to_csv(report_dir)
         self.create_images(report_dir)
         self.create_json_file(report_dir, timestamp)
+        self.create_report_pdf(os.path.join(report_dir, "report.json"))
 
     @staticmethod
     def get_report_dir_from_node(default_report_dir, node1, node2, timestamp):
@@ -307,3 +314,263 @@ class ReportCreationMixin:
         # write the report to a json file
         with open(report_path, "w") as f:
             json.dump(report_dict, f, indent=4)
+
+    def create_report_pdf(self, report_json_path):
+        """
+        Create a PDF report from the JSON file.
+
+        Args:
+            report_json_path (str): Path to the JSON file.
+        """
+
+        # Read the JSON file
+
+        with open(report_json_path, "r") as f:
+            report_data = json.load(f)
+         
+        styles = getSampleStyleSheet()
+
+        header_style = ParagraphStyle(
+            name="CenteredHeader",
+            fontSize=10,
+            textColor=colors.black,
+            alignment=TA_CENTER,
+            parent=styles['Normal'],
+        )
+
+        subtitle_style = ParagraphStyle(
+            name="Subtitle",
+            fontSize=14,
+            leading=18,
+            textColor=colors.black,
+            alignment=TA_LEFT,
+            spaceAfter=6
+        )
+
+        table_style_header = TableStyle([
+            ('GRID', (0, 0), (-1, -1), 0.5, colors.grey),
+            ('BACKGROUND', (0, 0), (-1, 0), colors.darkgrey),  # Color for header (row 0)
+            ('BACKGROUND', (0, 1), (-1, -1), colors.lightgrey),  # Color for all rows after the header (rows 1 to end)
+            ('VALIGN', (0, 0), (-1, -1), 'MIDDLE'),
+            ('ALIGN', (0, 0), (-1, -1), 'CENTER'),  # Horizontal alignment
+        ])
+
+        table_style_noheader = TableStyle([
+            ('GRID', (0, 0), (-1, -1), 0.5, colors.grey),
+            ('BACKGROUND', (0, 0), (-1, -1), colors.lightgrey),  # Color for all rows after the header (rows 1 to end)
+            ('VALIGN', (0, 0), (-1, -1), 'MIDDLE'),
+            ('ALIGN', (0, 0), (-1, -1), 'CENTER'),  # Horizontal alignment
+        ])
+
+        spacer = KeepTogether(Spacer(1, 0.25 * inch))
+        spacer2 = KeepTogether(Spacer(1, 0.5 * inch))
+         
+        # Set up document
+        doc = SimpleDocTemplate(report_json_path.replace(".json", ".pdf"), pagesize=letter)
+
+        # Get styles
[…]
