+++ docker/Dockerfile
@@ -81,6 +81,7 @@ RUN /home/researcher/slicer/Slicer-5.8.1-linux-amd64/bin/PythonSlicer -m pip ins
 RUN /home/researcher/slicer/Slicer-5.8.1-linux-amd64/bin/PythonSlicer -m pip install git+https://github.com/MIC-DKFZ/HD-BET@3e3d2a5993e8288f2eae928744ffb496cfc7c651
 RUN /home/researcher/slicer/Slicer-5.8.1-linux-amd64/bin/PythonSlicer -m pip install pyyaml==6.0.2
 RUN /home/researcher/slicer/Slicer-5.8.1-linux-amd64/bin/PythonSlicer -m pip install antspyx==0.5.4
+RUN /home/researcher/slicer/Slicer-5.8.1-linux-amd64/bin/PythonSlicer -m pip install reportlab==4.4.1
 
 # set home directory
 ENV HOME /home/researcher
