+++ docs/build/html/_downloads/f6975e4f3732f5bd835726bd4e26431b/run_tests_docker.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-THIS_DIR=$(dirname "$(readlink -f "$0")")
-${THIS_DIR}/../docker/docker_run.sh "/home/researcher/rano2.0-assist/tests/run_tests.sh"
