+++ docs/build/html/_downloads/5266e958c0d7fec22be7ee0718afab9e/test_rano.py
@@ -1,422 +1,0 @@
-import os
-import sys
-import time
-from glob import glob
-import json
-
-import qt
-import slicer
-from slicer.ScriptedLoadableModule import *
-from DICOMLib import DICOMUtils
-
-from utils.config import module_path, test_data_path
-
-
-#
-# RANOTest
-#
-class RANOTest(ScriptedLoadableModuleTest):
-    """
-    This is the test class for the RANO module.
-    It is used to test the RANO module and its functionality.
-    Upon clicking the "Reload and Test" button in the module (only shown when the module is loaded in developer mode),
-    this classes runTest() method is called.
-    """
-    def setUp(self):
-        """ Do whatever is needed to reset the state - typically a scene clear will be enough.
-        """
-        slicer.mrmlScene.Clear()
-
-    def runTest(self):
-        """Run as few or as many tests as needed here.
-        """
-        self.setUp()
-
-        slicer.mrmlScene.Clear()
-
-        # check which tests were specified in the command line
-        tests_to_run = []
-        for arg in sys.argv:
-            if arg.startswith("test_"):
-                tests_to_run.append(arg)
-
-        if len(tests_to_run) > 0:  # if tests were specified in the command line
-            for test in tests_to_run:
-                if hasattr(self, test):
-                    print(f"Running test {test}")
-                    # run the test
-                    try:
-                        getattr(self, test)()
-                    except Exception as e:
-                        print(f"Test {test} failed with error: {e}")
-                        sys.exit(1)
-                else:
-                    print(f"Test {test} not found")
-        else:  # if no tests were specified run all tests
-            self.test_RANO_dicom_KCL()
-            #self.test_RANO_dicom_CPTAC()
-            self.test_RANO_nifti_MU()
-
-        if any(['.py' in arg for arg in sys.argv]):
-            print('Tests executed from command line. Quitting Slicer...')
-            slicer.app.quit()
-
-    def test_RANO_dicom_KCL(self):
-        slicer.mrmlScene.Clear()
-        self.delayDisplay("Starting the test")
-
-        '''
-        define the test cases
-        '''
-
-        base_path = os.path.join(test_data_path, 'KCL')
-
-        patients = sorted(glob(os.path.join(base_path, "Patient_*")))
-
-        kcl_dcm_timepoint_pairs = [((patient, f"TimePoint_001"), (patient, f"TimePoint_002"))
-                                  for patient in patients]
-
-        cases_of_paths_2tp_times_4channels = []  # n cases, 2 timepoints, 4 channels
-        for test_case_idx in range(len(kcl_dcm_timepoint_pairs)):
-            patient = kcl_dcm_timepoint_pairs[test_case_idx][0][0]
-            timepoint_t1 = kcl_dcm_timepoint_pairs[test_case_idx][0][1]
-            timepoint_t2 = kcl_dcm_timepoint_pairs[test_case_idx][1][1]
-
-            p_t1c_tp1 = os.path.join(base_path, patient, timepoint_t1, "t1c")
-            p_t1n_tp1 = os.path.join(base_path, patient, timepoint_t1, "t1n")
-            p_t2f_tp1 = os.path.join(base_path, patient, timepoint_t1, "t2f")
-            p_t2w_tp1 = os.path.join(base_path, patient, timepoint_t1, "t2w")
-
-            p_t1c_tp2 = os.path.join(base_path, patient, timepoint_t2, "t1c")
-            p_t1n_tp2 = os.path.join(base_path, patient, timepoint_t2, "t1n")
[…]
