### 2025-05-28 ad72569
Tests now check additionally if a new report pdf file was created
