+++ tests/run_tests.sh
@@ -33,13 +33,48 @@ echo "Using Slicer executable: $SLICER_EXECUTABLE"
 
 THIS_DIR=$(dirname "$(readlink -f "$0")")
 
+REPORTS_DIR="${THIS_DIR}/../Reports"
+
+echo "Reports directory: $REPORTS_DIR"
+
 test_summary=""
 # run tests one by one
 for test in "${tests[@]}"; do
     echo "Running $test..."
+    # run the test script in Slicer
     $SLICER_EXECUTABLE --no-splash --python-script ${THIS_DIR}/run_tests.py "$test"
-    # Check if the Slicer executable ran successfully
-    if [ $? -eq 0 ]; then
+
+    # Get all report files in the Reports directory again after running the test
+    report_files=("${REPORTS_DIR}"/*/*.pdf)
+
+    # get the last modified date of all report files
+    report_dates=()
+    for file in "${report_files[@]}"; do
+        if [ -f "$file" ]; then
+            report_dates+=("$(stat -c %Y "$file")")
+        fi
+    done
+
+    current_time=$(date +%s)
+
+    # date differences
+    date_diffs=()
+    for date in "${report_dates[@]}"; do
+        diff=$((current_time - date))
+        date_diffs+=("$diff")
+    done
+
+    # sort the date differences
+    sorted_diffs=$(printf "%s\n" "${date_diffs[@]}" | sort -n)
+    min_diff=${sorted_diffs[0]}
+
+    # in case of no reports, set min_diff to a large number
+    if [ ${#report_dates[@]} -eq 0 ]; then
+        min_diff=9999999999
+    fi
+
+    # Check if the Slicer executable ran successfully and if the newest report was created within the last 10 seconds
+    if [ $? -eq 0 ] && [ "$min_diff" -lt 10 ]; then
         test_summary+="Test $test: PASSED\n"
     else
         test_summary+="Test $test: FAILED\n"
