### 2025-05-28 66c3424
Default to PythonSlicer and fixed tests
