### 2025-05-28 4a0cf62
Removed CLI status printout after segmentation because it reports errors when the subprocess reports warnings
