### 2025-05-28 1357f68
Set allow_smaller=False at during testing to avoid deprecation warning
