### 2025-05-28 2f35fe2
Supress ITK based error/warning that happens when antspyx uses python-real's ITK
