### 2025-05-28 8546768
Re-add test in run_tests.sh
