### 2025-09-27 #8246 @dstansby
@RuoyanMeng would you be interested in reviewing this JOSS submission? For more information on what reviewing for JOSS involves, see here: [joss.readthedocs.io/en/latest/reviewer_guidelines.html](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html), and a preview of the sumbitted article can be viewed here: [openjournals/joss-papers@joss.08246/joss.08246/10.21105.joss.08246.pdf](https://github.com/openjournals/joss-papers/blob/joss.08246/joss.08246/10.21105.joss.08246.pdf).
