### 2025-12-16 acae42f
Added pip_install commands to RANO.py
