### 2025-12-16 #9629 @aaronkujawa
Thanks again for the review/comments, @zapaishchykova.

Point 1, 4, 5 are clear, and I'll make the adjustments.

Point 3 seems to be a downstream issue of point 2, which seems to be a problem with the Zenodo directory merging:
The error message in point 2 says that the "dynunet_pipeline/data/tasks" directory cannot be found. In the documentation's [installation instructions](https://cai4cai.ml/rano2.0-assist/build/html/index.html#segmentation-models-and-test-data) , it says that this directory (among others) has to be added to the cloned repository. (The model weights are too large to be included on github). After merging, the directory should exist. Can you please double check that you downloaded and included the Zenodo directories as described? Please let me know if this solves this issue and whether the documentation needs clarification.
