### 2025-12-16 #9629 @zapaishchykova
if I had to guess, it has something to do with the way how it constructs the path (see screenshot in the issue 2)
