### 2025-12-17 #9629 @zapaishchykova
Ok thank you, keep me posted on the updates  - I will try to find a Windows/Linux system with GPU in the meantime
