### 2025-12-23 #9629 @aaronkujawa
Hi @zapaishchykova ,
Great, thank you for trying on Windows. I had only tested the repository on Linux, but will make the necessary adjustments for Windows next week. Sorry about the delay, I currently have no access to a Windows machine but will be able to access one next week.
