### 2025-12-30 be54bcd
Bugfix: ctkErrorLogLevel on Windows is "Unknown", not "None". Check for both options
