### 2025-12-30 8e4060c
Bugfix: Setting progressBar progress to 99 instead of 100, because it automatically closes at 100%
