### 2025-12-30 4a343f4
Bugfix: had imported ScriptedLoadableModule module instead of class
