### 2025-12-30 28295d9
Bugfix: Quering stdout and stderr in run_command.py led to blocking issue. Now instead using threading and queue to avoid blocking.
