+++ run_command/run_command.py
@@ -6,6 +6,8 @@ import subprocess
 import re
 from pathlib import Path
 from slicer.ScriptedLoadableModule import ScriptedLoadableModule
+import threading
+import queue
 
 if __name__ == '__main__':
     # define progress bar steps as fraction of 1
@@ -16,85 +18,110 @@ if __name__ == '__main__':
 
     command = sys.argv[1]
 
-    # send progress bar start line to stdout
+    # send progress bar start line
     print("""<filter-start><filter-name>TestFilter</filter-name><filter-comment>ibid</filter-comment></filter-start>""")
     sys.stdout.flush()
-
-    print("""<filter-progress>{}</filter-progress>""".format(command_received_progress))
+    print(f"<filter-progress>{command_received_progress}</filter-progress>")
     sys.stdout.flush()
 
-    # create startup environment for subprocess (to run python outside slicer)
+    # startup environment
     slicer_path = Path(os.environ["SLICER_HOME"]).resolve()
     PATH_without_slicer = os.pathsep.join(
-        [p for p in os.environ["PATH"].split(os.pathsep) if not slicer_path in Path(p).parents])
+        [p for p in os.environ["PATH"].split(os.pathsep) if not slicer_path in Path(p).parents]
+    )
+    startupEnv = {
+        "PATH": PATH_without_slicer,
+        "SYSTEMROOT": os.environ.get("SYSTEMROOT", ""),
+        "USERPROFILE": os.environ.get("USERPROFILE", "")
+    }
 
-    startupEnv = {}
-    try:
-        startupEnv["SYSTEMROOT"] = os.environ["SYSTEMROOT"]
-    except:
-        print("SYSTEMROOT environment variable does not exist...")
+    popen = subprocess.Popen(
+        command,
+        stdout=subprocess.PIPE,
+        stderr=subprocess.PIPE,
+        universal_newlines=True,
+        env=startupEnv,
+        shell=True
+    )
 
-    try:
-        startupEnv["USERPROFILE"] = os.environ["USERPROFILE"]
-    except:
-        print("USERPROFILE environment variable does not exist...")
-
-    startupEnv['PATH'] = PATH_without_slicer
-
-    popen = subprocess.Popen(command,
-                             stdout=subprocess.PIPE,
-                             stderr=subprocess.PIPE,
-                             universal_newlines=True,
-                             env=startupEnv,
-                             shell=True)
     stdout_all = ""
     stderr_all = ""
-    while True:
+    progress_queue = queue.Queue()
+
+
+    # Reader for stdout
+    def read_stdout(pipe):
+        nonlocal_stdout = []
         loop_idx = 0
-        stdout_line = popen.stdout.readline()
-        stderr_line = popen.stderr.readline()
+        for line in iter(pipe.readline, ''):
+            #print("STDOUT:", line.strip())
+            if loop_idx == 0:
+                # enqueue initial progress update
+                progress_queue.put(first_output_progress)
+            if line:
+                nonlocal_stdout.append("run_command stdout: " + line)
+            loop_idx += 1
+        pipe.close()
+        # merge back into global
+        nonlocal_stdout_str = "".join(nonlocal_stdout)
+        if nonlocal_stdout_str:
+            progress_queue.put(("stdout_all", nonlocal_stdout_str))
 
-        if loop_idx == 0:
-            print("""<filter-progress>{}</filter-progress>""".format(first_output_progress))
-            sys.stdout.flush()
 
-        if stdout_line:
-            stdout_all += "run_command stdout: " + stdout_line
+    # Reader for stderr
+    def read_stderr(pipe):
+        nonlocal_stderr = []
+        for line in iter(pipe.readline, ''):
+            #print("STDERR:", line.strip())
[…]
