### 2025-12-30 4968fcf
Make sure stdout is not modified by imports. (new version of HD-BET does that, subsequently no output in slicer python console)
