### 2025-12-30 5cbf434
avoid using Linux tee by custom python Tee class
