### 2025-12-31 972c5a2
When debug=True, wrong line pairs were calculated (for speedup). This was fixed now.
