### 2025-12-31 3a52442
update hd-bet version
