+++ docs/build/html/_sources/README.md.txt
@@ -20,7 +20,7 @@ the following steps:
 # Installation
 
 ## System requirements
-- **Operating System**: The tool has been tested on Ubuntu 24.04
+- **Operating System**: The tool has been tested on Ubuntu 24.04 and Windows 11
 - **GPU**: The tool requires a GPU with CUDA support for the automatic segmentation. The tool has been tested on
   NVIDIA GPUs with CUDA 12.4.
 
@@ -122,67 +122,53 @@ generated by the tool.
 Follow the links to install the required software:
 - [3D Slicer](https://download.slicer.org/) (tested with version 5.8.1)
 
-The following Python packages can be installed in 3D Slicer's python environment. 
-The versions listed below are the ones that have been tested with this extension.
-To install the packages, you can either use the pip_install command in the 3D Slicer Python Console (e.g. 
-type `pip_install("numpy==2.0.2")` in the console) or use the command line to install the packages in the 3D Slicer Python
-environment as follow:
 
-- numpy (tested with version 2.0.2): `pip_install("numpy==2.0.2")`
-- scikit-image (tested with version 0.24.0): `pip_install("scikit-image==0.24.0")`
-- numba (tested with version 0.60.0): `pip_install("numba==0.60.0")`
-- nibabel (tested with version 5.3.2): `pip_install("nibabel==5.3.2")`
-- tqdm (tested with version 4.67.1) : `pip_install("tqdm==4.67.1")`
-- pyyaml (tested with version 6.0.2): `pip_install("pyyaml==6.0.2")`
-- reportlab (tested with version 4.4.1): `pip_install("reportlab==4.4.1")`
+#### PyTorch Extension
+Next, the PyTorch extension needs to be installed within 3D Slicer. This
+extension facilitates the installation of PyTorch in 3D Slicer.
+- Open 3D Slicer
+- Open the Extension Manager: "View" --> "Extension Manager" 
+- Install the PyTorch extension: "Install Extensions" --> "Utilities" --> "PyTorch"
+- Restart 3D Slicer
 
-The following packages are required for the segmentation pipeline (DynUNet). These can be installed either in the 3D Slicer
-python environment or in a separate python environment whose path can be specified in the RANO module. The following 
-commands are to install the packages in the 3D Slicer python environment.
-- PyTorch: check here which pip command is appropriate for your system (https://pytorch.org/get-started/locally/) (tested with version 2.6)
-    - for example, in 3D Slicer's python console type: `pip_install("torch==2.6.0+cu124 torchvision==0.21.0+cu124 torchaudio==2.6.0+cu124 --index-url https://download.pytorch.org/whl/cu124")`
-- MONAI (install from https://github.com/aaronkujawa/MONAI/tree/rano): `pip_install("git+https://github.com/aaronkujawa/MONAI.git@rano")`
-- pytorch-ignite (tested with version pytorch-ignite 0.5.2): `pip_install("pytorch-ignite==0.5.2")`
-- tensorboard (tested with version 2.19.0): `pip_install("tensorboard==2.19.0")`
-- ANTsPy (tested with version 0.5.4): `pip_install("antspyx==0.5.4")`
-- HD-BET (requires this specific commit): `pip_install("git+https://github.com/MIC-DKFZ/HD-BET@3e3d2a5993e8288f2eae928744ffb496cfc7c651")`
-
-Note: if a separate python environment is used, numpy, scikit-image, numba, nibabel, tqdm and pyyaml need to be installed
-in that environment on top of the installation in 3D Slicer.
-
-Summary of commands for the 3D Slicer Python Console:
[…]
