### 2026-01-03 2a186c4
Improved automated dependency installations
