### 2026-01-03 5736a61
Install ants form wheel provided in SlicerAntsPy Extension
