### 2026-01-03 865eadd
restrict pytorch version to 2.8.0 because 2.9 reportedly causes OOM issues
