+++ RANO/utils/report_creation_utils.py
@@ -25,15 +25,6 @@ class ReportCreationMixin:
     """
     Mixin class for creating a report for the RANO module.
     """
-    def __init__(self, parameterNode, ui, lineNodePairs):
-        self._parameterNode = parameterNode
-        """Parameter node for the RANO module"""
-
-        self.ui = ui
-        """UI for the RANO module"""
-
-        self.lineNodePairs = lineNodePairs
-        """List of line node pairs for the RANO module"""
 
     def onCreateReportButton(self):
         """
