### 2026-01-07 740ccd7
Delayed dependency installation (at module opening) and dynamical addition of Mixins
