+++ RANO/RANO.py
@@ -13,119 +13,27 @@ from importlib import reload
 import slicer, vtk
 from slicer.ScriptedLoadableModule import *
 from slicer.util import *
-slicer_stdout = sys.stdout
 
-try:
-    import PyTorchUtils
-except ModuleNotFoundError as e:
-    raise Exception("This module requires PyTorch extension. Install it from the Extensions Manager.")
-
-def ensure_package(import_name, spec, verbose=True):
-    import importlib
-
-    try:
-        if verbose:
-            print(f"[ensure_package] importing {import_name}")
-        return importlib.import_module(import_name)
-    except ModuleNotFoundError:
-        if verbose:
-            print(f"[ensure_package] installing {spec}")
-        pip_install(spec)
-        importlib.invalidate_caches()
-        return importlib.import_module(import_name)
-
-numpy = ensure_package("numpy", "numpy==2.0.2")
-skimage = ensure_package("skimage", "scikit-image==0.24.0")
-numba = ensure_package("numba", "numba==0.60.0")
-nibabel = ensure_package("nibabel", "nibabel==5.3.2")
-tqdm = ensure_package("tqdm", "tqdm==4.67.1")
-yaml = ensure_package("yaml", "pyyaml==6.0.2")
-reportlab = ensure_package("reportlab", "reportlab==4.4.1")
-
-try:
-    import torch
-except:
-    minimumTorchVersion = "2.6.0"
-    maximumTorchVersion = "2.8.0"
-
-    torchRequirement = f">={minimumTorchVersion},<={maximumTorchVersion}"
-
-    torchLogic = PyTorchUtils.PyTorchUtilsLogic()
-    if not torchLogic.torchInstalled():
-        print('PyTorch Python package is required. Installing... (it may take several minutes)')
-        torch = torchLogic.installTorch(askConfirmation=False, torchVersionRequirement=torchRequirement)
-        if torch is None:
-            raise Exception("This module requires PyTorch extension. Install it from the Extensions Manager.")
-    import torch
-
-ensure_package("ignite", "pytorch-ignite==0.5.2")
-ensure_package("tensorboard", "tensorboard==2.19.0")
-
-def installANTsPyX():
-    try:
-        import ants
-    except:
-        import platform
-        if platform.system() == 'Linux':
-            import urllib.request
-            import tempfile
-
-            # Download the wheel file from Box
-            download_url = 'https://app.box.com/shared/static/mu1gy26t80oopbtv3mndl5yveb6s4431.whl'
-            temp_dir = tempfile.gettempdir()
-            whl_filename = 'antspyx-0.6.2-cp312-cp312-linux_x86_64.whl'
-            whl_path = os.path.join(temp_dir, whl_filename)
-
-            print(f"Downloading antspyx wheel from {download_url}")
-            urllib.request.urlretrieve(download_url, whl_path)
-            print(f"Downloaded to {whl_path}")
-
-            slicer.util.pip_install(whl_path)
-
-            # Clean up the downloaded file
-            try:
-                os.remove(whl_path)
-            except:
-                pass
-        else:
-            slicer.util.pip_install('antspyx')
-
-installANTsPyX()
-
-ensure_package("HD_BET", "hd-bet==2.0.1")
-ensure_package("monai", "git+https://github.com/aaronkujawa/MONAI.git@rano")
-
-from utils.config import debug
-sys.stdout = slicer_stdout
-
-# reload the submodules to ensure that the latest version of the code is used upon pressing the "Reload" button in the
-# module GUI (development mode)
-for mod in ['utils.enums',  # must be first to ensure that the subsequent modules are reloaded with the new enums
-            'utils.config',
-            'utils.rano_utils',
-            'utils.RANO_Logic',
-            'utils.ui_helper_utils',
-            'utils.segmentation_utils',
[…]
