### 2026-01-07 #9629 @dstansby
Hi @aaronkujawa, just checking in and wanted to see how fixing this on Windows was going? I'll consider the review on hold until you get back to us, but I appreciated it's been winter holiday season so you might not have had time in the last couple of weeks.
