### 2026-01-07 9ca8682
Added installation of antspyx library
