+++ RANO/RANO.py
@@ -438,6 +438,42 @@ def installAndImportDependencies():
     ensure_package("tensorboard", "tensorboard==2.19.0")
 
     ensure_package("HD_BET", "hd-bet==2.0.1")
+
+    def installANTsPyX():
+        """
+        source: https://github.com/SlicerMorph/SlicerANTsPy/blob/main/ANTsRegistration/ANTsRegistration.py
+        """
+        try:
+            import ants
+        except:
+            import platform
+            if platform.system() == 'Linux':
+                import urllib.request
+                import tempfile
+
+                # Download the wheel file from Box
+                download_url = 'https://app.box.com/shared/static/mu1gy26t80oopbtv3mndl5yveb6s4431.whl'
+                temp_dir = tempfile.gettempdir()
+                whl_filename = 'antspyx-0.6.2-cp312-cp312-linux_x86_64.whl'
+                whl_path = os.path.join(temp_dir, whl_filename)
+
+                print(f"Downloading antspyx wheel from {download_url}")
+                urllib.request.urlretrieve(download_url, whl_path)
+                print(f"Downloaded to {whl_path}")
+
+                pip_install(whl_path)
+
+                # Clean up the downloaded file
+                try:
+                    os.remove(whl_path)
+                except:
+                    pass
+            else:
+                pip_install('antspyx')
+
+    installANTsPyX()
+    import ants
+
     ensure_package("monai", "git+https://github.com/aaronkujawa/MONAI.git@rano")
 
     sys.stdout = slicer_stdout
