+++ docs/build/html/_sources/README.md.txt
@@ -120,7 +120,7 @@ generated by the tool.
 
 #### Software requirements
 Follow the links to install the required software:
-- [3D Slicer](https://download.slicer.org/) (tested with version 5.8.1)
+- [3D Slicer](https://download.slicer.org/) (tested with version 5.10.0)
 
 
 #### PyTorch Extension
@@ -147,10 +147,10 @@ To add the extension to 3D Slicer, follow these steps:
 
 <img src="_static/confirm_module_import.png" alt="drawing" width="200"/>
 
-1. Loading the module may take several minutes, since it installs various Python libraries (listed below).
-
 1. Start the RANO module by selecting "Tools" --> "RANO" in the Modules drop down menu or search for
    "RANO" using the search tool.
+1. Opening the module for the first time may take several minutes, since it installs various Python libraries (listed below) automatically.
+
 
 # Python libraries
 The following libraries will be installed when the RANO2.0-assist module is first loaded:
