### 2026-01-09 #9629 @zapaishchykova
Hello @aaronkujawa sorry my bad, somehow missed it, but fixed it now. Please see the new error log below:

```
Running 1 test cases
Starting the automatic segmentation
[VTK] run_command standard error:
[VTK] Traceback (most recent call last):
[VTK]   File "E:\Anna\Rano\rano2.0-assist-main\run_command\run_command.py", line 124, in <module>
[VTK]     raise subprocess.CalledProcessError(popen.returncode, command)
[VTK] subprocess.CalledProcessError: Command '"C:/Users/jlzie/AppData/Local/slicer.org/Slicer 5.6.1/bin/PythonSlicer.exe" "E:\Anna\Rano\rano2.0-assist-main\RANO\..\dynunet_pipeline\src\inference.py" -task_dir "E:\Anna\Rano\rano2.0-assist-main\dynunet_pipeline\data\tasks\task4001_brats24allmod" -args_file "config/infer_args.json" -inference_files_path "E:\Anna\Rano\rano2.0-assist-main\RANO\timepoint1\tmp_in\inference_files.json" -out_dir "E:\Anna\Rano\rano2.0-assist-main\RANO\timepoint1\tmp_out" -reg  -no-input_is_bet ' returned non-zero exit status 1.
[VTK] run_command completed with errors
Segmentation for timepoint1 completed after 11 seconds.
[Python] Output segmentation file not found: E:\Anna\Rano\rano2.0-assist-main\RANO\timepoint1\tmp_out\preprocessed\registered\output.nii.gz
[VTK] run_command standard error:
[VTK] Traceback (most recent call last):
[VTK]   File "E:\Anna\Rano\rano2.0-assist-main\run_command\run_command.py", line 124, in <module>
[VTK]     raise subprocess.CalledProcessError(popen.returncode, command)
[VTK] subprocess.CalledProcessError: Command '"C:/Users/jlzie/AppData/Local/slicer.org/Slicer 5.6.1/bin/PythonSlicer.exe" "E:\Anna\Rano\rano2.0-assist-main\RANO\..\dynunet_pipeline\src\inference.py" -task_dir "E:\Anna\Rano\rano2.0-assist-main\dynunet_pipeline\data\tasks\task4001_brats24allmod" -args_file "config/infer_args.json" -inference_files_path "E:\Anna\Rano\rano2.0-assist-main\RANO\timepoint2\tmp_in\inference_files.json" -out_dir "E:\Anna\Rano\rano2.0-assist-main\RANO\timepoint2\tmp_out" -reg  -no-input_is_bet ' returned non-zero exit status 1.
[VTK] run_command completed with errors
Segmentation for timepoint2 completed after 5 seconds.
[Python] Output segmentation file not found: E:\Anna\Rano\rano2.0-assist-main\RANO\timepoint2\tmp_out\preprocessed\registered\output.nii.gz

```
