### 2026-01-11 96cab55
Added community guidelines to docs
