### 2026-01-11 0e57fef
Removed test cases box from UI and setup_test_cases method
