### 2026-01-11 b8210d3
Added State of the field section to manuscript
