### 2026-01-11 a7adf0c
Added State of the field section to manuscript
