+++ JOSS/paper.bib
@@ -1,3 +1,20 @@
+@misc{mintlesion
+  author       = {{Mint Medical GmbH}},
+  title        = {Implementing RANO 2.0 for Neuro-Oncology Clinical Trials in mint Lesion},
+  year         = {2025},
+  month        = {Nov},
+  howpublished = {\url{https://mint-medical.com/news/implementing-rano-20-for-neuro-oncology-clinical-trials-in-mint-lesion}},
+  note         = {Accessed: 2026-01-11},
+}
+
+@misc{graylight_rano,
+  author       = {{Graylight Imaging Sp. z o.o.}},
+  title        = {Automated RANO solution for clinical trial},
+  year         = {2026},
+  howpublished = {\url{https://cdn.graylight-imaging.com/project/automated-rano-solution-for-clinical-trial/}},
+  note         = {Accessed: 2026-01-11},
+}
+
 @article{fedorov20123d,
   title={3D Slicer as an image computing platform for the Quantitative Imaging Network},
   author={Fedorov, Andriy and Beichel, Reinhard and Kalpathy-Cramer, Jayashree and Finet, Julien and Fillion-Robin, Jean-Christophe and Pujol, Sonia and Bauer, Christian and Jennings, Dominique and Fennessy, Fiona and Sonka, Milan and others},
