### 2026-01-23 #9629 @dstansby
Hi @zapaishchykova thanks for your time spent reviewing this!

Since I've struggled to find a second reviewer, and I work myself in medical imaging, I've decided to review this myself as the second reviewer. @aaronkujawa hopefully this won't take more than a couple of weeks, apologies for the delay.
