### 2026-01-24 #9629 @aaronkujawa
Hi @dstansby , thank you for the offer to review.
