### 2026-01-30 #9629 @dstansby
- [x] **License:** Does the repository contain a plain-text LICENSE or COPYING file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license? (but would be good to fix https://github.com/cai4cai/rano2.0-assist/issues/4)

- [x] **Installation:** Does installation proceed as outlined in the documentation? (https://github.com/cai4cai/rano2.0-assist/issues/7)

- [x] **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution. (https://github.com/cai4cai/rano2.0-assist/issues/6)
- [x] **Automated tests:** Are there automated tests or manual steps described so that the functionality of the software can be verified? (https://github.com/cai4cai/rano2.0-assist/issues/8)
