### 2026-01-30 cai4cai/rano2.0-assist#4 @dstansby
Add year/copyright owner to README
I noticed here in the README:

https://github.com/cai4cai/rano2.0-assist/blob/b8210d39c5b6329381d9af81b23e34df847e8ad7/LICENSE#L189

The copyright owner and year need filling in.
