### 2026-01-30 cai4cai/rano2.0-assist#6 @dstansby
Note that docker installation only works on Linux
I tried following the docker install instructions, but it seems like the NVIDIA container toolkit can only be installed on Linux. It would be good to mention that the docker install instructions only work on Linux in the docs here.
