### 2026-01-30 cai4cai/rano2.0-assist#7 @dstansby
Explain where to find Slicer path for running tests
To run the tests, the documentation says that Slicer binary has to be in the path, but it's not obvious to me:

1. How to find the location of the slicer binary
2. Add this to the path

Both points should be explained in the documentation.
