### 2026-01-30 cai4cai/rano2.0-assist#8 @dstansby
Tests don't run when there's a space in the slicer path
When there's a space in the slicer path (I have a folder named "David S" in the path), the tests fail. I've checked that putting quotes around $SLICER_EXECUTABLE on this line fixes it:

https://github.com/cai4cai/rano2.0-assist/blob/b8210d39c5b6329381d9af81b23e34df847e8ad7/tests/run_tests.sh#L43
