### 2026-02-01 cai4cai/rano2.0-assist#4 @aaronkujawa
Thanks, good catch. These were added.
