### 2026-02-03 3ab0eec
Added automatic download of model weights and test data from Zenodo
