### 2026-02-07 7c430b2
bugfix: run_tests.sh was not working with Slicer on PATH
