### 2026-02-08 250187b
made x11 work in docker to install extension via dockerfile
