+++ docker/install_dependencies.py
@@ -1,0 +1,4 @@
+import sys
+sys.path.append('/home/researcher/rano2.0-assist/RANO')
+from RANO import installAndImportDependencies
+installAndImportDependencies()
