+++ docker/Dockerfile
@@ -4,6 +4,9 @@ FROM nvcr.io/nvidia/pytorch:25.03-py3
 RUN SLICER_URL="https://download.slicer.org/bitstream/67c51fc129825655577cfee9" && \
   curl -k -v -s -L $SLICER_URL | tar xz -C /tmp
 
+
+# https://download.slicer.org/bitstream/6911b598ac7b1c95e7934427  # slicer 5.10.0
+
 # Install GUI testing tools
 RUN apt-get update && \
     apt-get install -y libglu1 libpulse-mainloop-glib0 libnss3 libasound2t64 qt5dxcb-plugin  #
@@ -20,7 +23,20 @@ RUN dpkg --add-architecture i386 && \
         libxdmcp6 libxdmcp6:i386 \
         libxcb1 libxcb1:i386 \
         libxext6 libxext6:i386 \
-        libx11-6 libx11-6:i386 && \
+        libx11-6 libx11-6:i386 \
+        vim net-tools curl \
+         xserver-xorg-video-dummy \
+         libxrender1 \
+         libpulse0 \
+         libpulse-mainloop-glib0  \
+         libnss3  \
+         libxcomposite1 \
+         libxcursor1 \
+         libfontconfig1 \
+         libxrandr2 \
+         libglu1 \
+         awesome \
+        x11vnc && \
     rm -rf /var/lib/apt/lists/*
 
 # nvidia-container-runtime
@@ -65,23 +81,15 @@ RUN touch /etc/pip/constraint.txt
 
 RUN apt-get update && apt-get install -y --no-install-recommends xvfb
 
+
 # set the user
 USER researcher
-
-# install pip packages in Slicer
 RUN /home/researcher/slicer/Slicer-5.8.1-linux-amd64/bin/PythonSlicer -m pip install numpy==2.0.2
 RUN /home/researcher/slicer/Slicer-5.8.1-linux-amd64/bin/PythonSlicer -m pip install scikit-image==0.24.0
 RUN /home/researcher/slicer/Slicer-5.8.1-linux-amd64/bin/PythonSlicer -m pip install nibabel==5.3.2
 RUN /home/researcher/slicer/Slicer-5.8.1-linux-amd64/bin/PythonSlicer -m pip install tqdm==4.67.1
 RUN /home/researcher/slicer/Slicer-5.8.1-linux-amd64/bin/PythonSlicer -m pip install numba==0.60.0
 RUN /home/researcher/slicer/Slicer-5.8.1-linux-amd64/bin/PythonSlicer -m pip install torch==2.6.0+cu124 torchvision==0.21.0+cu124 torchaudio==2.6.0+cu124 --index-url https://download.pytorch.org/whl/cu124
-RUN /home/researcher/slicer/Slicer-5.8.1-linux-amd64/bin/PythonSlicer -m pip install tensorboard==2.19.0
-RUN /home/researcher/slicer/Slicer-5.8.1-linux-amd64/bin/PythonSlicer -m pip install pytorch-ignite==0.5.2
-RUN /home/researcher/slicer/Slicer-5.8.1-linux-amd64/bin/PythonSlicer -m pip install git+https://github.com/aaronkujawa/MONAI.git@rano
-RUN /home/researcher/slicer/Slicer-5.8.1-linux-amd64/bin/PythonSlicer -m pip install hd-bet==2.0.1
-RUN /home/researcher/slicer/Slicer-5.8.1-linux-amd64/bin/PythonSlicer -m pip install pyyaml==6.0.2
-RUN /home/researcher/slicer/Slicer-5.8.1-linux-amd64/bin/PythonSlicer -m pip install antspyx==0.5.4
[…]
