### 2026-02-15 4952c3d
Updated Dockerfile to Slicer 5.10
