### 2026-02-15 4f7cfd0
Updated docs. How to add Slicer to PATH and NVIDIA container toolkit restriction to Linux
