### 2026-02-18 #9629 @dstansby
I've reviewed the docs and paper, and all looks good. For testing the functionality, I'm trying to load some test cases, but can't find them. I opened an issue at https://github.com/cai4cai/rano2.0-assist/issues/10.
