### 2026-02-22 7f62feb
Test case selection from UI possible again
