### 2026-02-23 #9629 @aaronkujawa
Hi @dstansby, you are right, I had removed the box, but it is now back.
