### 2026-02-24 #9629 @dstansby
@aaronkujawa thanks a lot for responding to all my comments - I'm happy now as reviewer, and as editor delighted to take this forward as an accepted paper. I need to sort out a few technical things, but will let you know the next steps shortly.
