### 2026-02-27 #9629 @dstansby
@aaronkujawa please could you update your `.bib` file to include DOIs for all articles that have them?
