### 2026-03-01 76d7d3e
Added DOIs to paper.bib
