+++ JOSS/paper.bib
@@ -23,7 +23,8 @@
   number={9},
   pages={1323--1341},
   year={2012},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.mri.2012.05.001}
 }
 
 @article{wen2023rano,
@@ -34,7 +35,8 @@
   number={33},
   pages={5187--5199},
   year={2023},
-  publisher={Wolters Kluwer Health}
+  publisher={Wolters Kluwer Health},
+  doi={10.1200/JCO.23.01059}
 }
 
 @article{wen2010updated,
@@ -45,7 +47,8 @@
   number={11},
   pages={1963--1972},
   year={2010},
-  publisher={American Society of Clinical Oncology}
+  publisher={American Society of Clinical Oncology},
+  doi={10.1200/JCO.2009.26.3541}
 }
 
 @article{chang2019automatic,
@@ -56,7 +59,8 @@
   number={11},
   pages={1412--1422},
   year={2019},
-  publisher={Oxford University Press US}
+  publisher={Oxford University Press US},
+  doi={10.1093/neuonc/noz106}
 }
 
 @article{nalepa2023deep,
@@ -66,7 +70,8 @@
   volume={154},
   pages={106603},
   year={2023},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.compbiomed.2023.106603}
 }
 
 @article{shi2025automated,
@@ -75,12 +80,14 @@
   journal={Neuro-oncology},
   pages={noaf037},
   year={2025},
-  publisher={Oxford University Press US}
+  publisher={Oxford University Press US},
+  doi={10.1093/neuonc/noaf037}
 }
 
 @article{de20242024,
   title={The 2024 Brain Tumor Segmentation (BraTS) challenge: glioma segmentation on post-treatment MRI},
   author={de Verdier, Maria Correia and Saluja, Rachit and Gagnon, Louis and LaBella, Dominic and Baid, Ujjwall and Tahon, Nourel Hoda and Foltyn-Dumitru, Martha and Zhang, Jikai and Alafif, Maram and Baig, Saif and others},
   journal={arXiv preprint arXiv:2405.18368},
-  year={2024}
+  year={2024},
+  doi={10.48550/arXiv.2405.18368}
 }
