### 2026-03-01 #9629 @aaronkujawa
Thank you @dstansby  for reviewing and accepting the paper! I added the DOIs now.
