### 2026-03-01 #9629 @aaronkujawa
I created a release on GitHub and am currently waiting for approval from the repository owner (cai4cai) to grant access to Zenodo for GitHub integration for automatic archiving. I'll let you know when the archive has been created.
