### 2026-03-02 74b691a
Enhance README with detailed project description
Expanded the README to provide detailed information about the RANO2.0-assist tool, including its features and functionality. Also added a badge with link to JOSS paper/review
