### 2026-03-02 9026473
Minor copy edits to JOSS paper
