### 2026-03-02 #9629 @dstansby
Brilliant, let me know when you have a DOI. Although it's a a bit more effort, you could always manually upload the software to Zenodo if you are having trouble getting the repository owner to grant access.

I've copy edited the paper at https://github.com/cai4cai/rano2.0-assist/pull/11, please take a look there and either merge or let me know if something looks off.
