### 2026-03-03 #9629 @aaronkujawa
Thank you for the edits. I merged them and created one more edit (correct name for Mark MacDonald). I created release v1.0.1. The corresponding DOI on Zenodo is [10.5281/zenodo.18853917](https://doi.org/10.5281/zenodo.18853917).
