### 2026-03-04 #9629 @arfon
☝️ looks to be a transient issue with the GitHub API @dstansby so worth trying again.
