### 2026-03-04 #9629 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
