### 2026-03-15 #9629 @Kevin-Mattheus-Moerman
@aaronkujawa As AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks, some of which require your attention:

👉 Your project appears to lack contributing guidelines at the moment. Please work to add these and to mention/link to them in your README. You could for instance create a `CONTRIBUTING.md` file (see here for some examples: https://contributing.md/example/).

👉 Please ensure the archive title matches the paper title

- [x] Copyedit pass/checked for typos
👉 `respons` should read `response`
