### 2026-03-15 dccccbd
Fixed typo and referred to community guidelines in README
