### 2026-03-15 #9629 @aaronkujawa
Thank you, @Kevin-Mattheus-Moerman .
I added a mention of the contributing guidelines in the README, corrected the archive title and authors, and fixed the typo in the PDF.
