### 2026-03-17 b36ec82
Create CITATION.cff with citation details
Added citation information for the software and authors.
